services/fraud-engine/requirements.txt
fastapi
uvicorn[standard]
celery
redis
pillow
imagehash
pytesseract
pymupdf
scikit-learn
numpy
pandas
transformers
torch
torchvision
python-multipart
aiofiles
boto3
networkx
psycopg2-binary
sqlalchemy
piexif
python-jose[cryptography]
prometheus-client
structlog
python-dotenv
